requirements.txt
Flask
python-dotenv
pydrive
pillow
requests
opencv-python
numpy